server/requirements.txt
sanic
sanic_ext
jsonschema 
httpx
python-decouple
##py -m pip install -r requirements.txt 